refactor(types): rename RethChunkProvider → ChunkStorageProvider, relocate MockChunkProvider (Phase 2)
Rename the trait to reflect its actual role (storage backend for PdService, not Reth/EVM),
delete MockChunkProvider from irys-types and move it into pd_service's test module,
and remove unnecessary test-utils feature flags from chain and reth-node-bridge Cargo.toml.
