+++ crates/chain/Cargo.toml
@@ -18,7 +18,7 @@ irys-database.workspace = true
 irys-domain.workspace = true
 irys-storage.workspace = true
 irys-reth-node-bridge.workspace = true
-irys-types = { workspace = true, features = ["test-utils"] }
+irys-types.workspace = true
 irys-api-server.workspace = true
 irys-config.workspace = true
 irys-testing-utils.workspace = true
