+++ crates/reth-node-bridge/Cargo.toml
@@ -17,7 +17,7 @@ reth-db.workspace = true
 reth-chainspec.workspace = true
 reth-node-builder.workspace = true
 reth-provider.workspace = true
-irys-reth = { workspace = true, features = ["test-utils"] }
+irys-reth.workspace = true
 reth-node-ethereum.workspace = true
 reth-e2e-test-utils.workspace = true
 alloy-rpc-types-engine.workspace = true
