+++ crates/types/src/chunk_provider.rs
@@ -29,8 +29,8 @@ impl ChunkConfig {
 }
 
 /// Provides unpacked chunks to PD precompile.
-/// Used as storage backend by PdChunkManager.
-pub trait RethChunkProvider: Send + Sync + std::fmt::Debug {
+/// Used as storage backend by PdService.
+pub trait ChunkStorageProvider: Send + Sync + std::fmt::Debug {
     /// Returns unpacked chunk bytes or `None` if not found.
     fn get_unpacked_chunk_by_ledger_offset(
         &self,
@@ -76,53 +76,3 @@ pub type PdChunkReceiver = mpsc::UnboundedReceiver<PdChunkMessage>;
 /// PdService populates this during provisioning; the PD precompile reads it directly.
 /// Keys are `(ledger: u32, offset: u64)` tuples matching `ChunkKey` semantics.
 pub type ChunkDataIndex = Arc<DashMap<(u32, u64), Arc<Bytes>>>;
-
-/// Mock chunk provider that returns zero-filled chunks.
-#[cfg(any(test, feature = "test-utils"))]
-#[derive(Debug, Clone)]
-pub struct MockChunkProvider {
-    config: ChunkConfig,
-    cached_chunk: Bytes,
-}
-
-#[cfg(any(test, feature = "test-utils"))]
-impl MockChunkProvider {
-    #[inline]
-    pub fn new() -> Self {
-        let config = ChunkConfig {
-            num_chunks_in_partition: 100,
-            chunk_size: 256_000,
-            entropy_packing_iterations: 0,
-            chain_id: 1,
-        };
-        let cached_chunk = Bytes::from(vec![0_u8; config.chunk_size as usize]);
-        Self {
-            config,
-            cached_chunk,
-        }
-    }
-}
-
-#[cfg(any(test, feature = "test-utils"))]
-impl Default for MockChunkProvider {
-    #[inline]
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-#[cfg(any(test, feature = "test-utils"))]
-impl RethChunkProvider for MockChunkProvider {
-    fn get_unpacked_chunk_by_ledger_offset(
-        &self,
-        _ledger: u32,
-        _ledger_offset: u64,
-    ) -> eyre::Result<Option<Bytes>> {
-        Ok(Some(self.cached_chunk.clone()))
-    }
-
-    #[inline]
-    fn config(&self) -> ChunkConfig {
-        self.config
-    }
-}
